+++ CHANGES.rst
@@ -6,6 +6,10 @@ Changelog
 ------------------
 
 - Add a validation step "Are you sure?" before launching items and meetings local roles update from the action button on portal_plonemeeting
+- Fixed ftw.labels :
+  - Jar storage that was a dict instead a PersistentMapping and that was making changes done to it not persisted;
+  - Go back to the 'data' tab on the MeetingConfig while removing a label from the labels portlet;
+  - Invalidate the ftw.labels faceted vocabulary when a label is added/updated/removed.
 
 4.1.8 (2019-10-14)
 ------------------